+++ NEWS.md
@@ -33,6 +33,7 @@ New features (Analysis):
 + Emit `PhanTypeMismatchArgumentInternalProbablyReal` instead of `PhanTypeMismatchArgumentInternal` in a few more cases.
 + Be stricter about checking if callables/closures have anything in common with other types.
 + Preserve more specific phpdoc types when the php 8.0 `mixed` type is part of the real type set.
++ Also emit `PhanPluginUseReturnValueNoopVoid` when a function/method's return type is implicitly void (#4049)
 
 Plugins:
 + Warn and skip checks instead of crashing when running `InlineHTMLPlugin` without the `tokenizer` extension installed. (#3998)